Split CLAP ABI callbacks
Move entry, plugin, and test callback wiring into focused modules so the adapter no longer needs a file-length quality exception.
